Refactor ComprehensiveValidator to follow SOLID principles
Based on PR review feedback, completely refactored the monolithic
ComprehensiveValidator class into a modular architecture:

- Created ValidationOrchestrator to coordinate all validators
- Split validation logic into focused validator classes:
  - SchemaValidator: YAML parsing and JSON schema validation
  - DependencyValidator: Tool, step reference, and resource validation
  - LintingValidator: Best practices and code quality checks
- Added StepCollector for efficient caching of step traversal
- Fixed all edge cases identified in review:
  - Proper nil/empty content handling
  - Regex match nil checks
  - Method naming (has_error_handling? -> error_handling?)
- Made validation thresholds configurable constants
- Removed incomplete security check feature
- ComprehensiveValidator now acts as a thin wrapper for backwards compatibility

This refactoring addresses all SOLID principle violations, improves
performance through caching, and creates a more maintainable and
extensible validation system.
